aws_vianova/requirements.txt
pulumi>=3.0.0,<4.0.0
pulumi-aws>=4.0.0,<5.0.0
python-dotenv==0.17.1

